pyproject.toml
[tool.poetry]
name = "ai-hitl-iso-compliance-fastapi"
version = "0.1.0"
description = ""
authors = ["imjwang <ngjaywa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "0.19.0"
copilotkit = "^0.1.25"
google-generativeai = "^0.8.3"
google-cloud-storage = "^2.18.2"
python-multipart = "^0.0.17"

weave = "^0.51.18"
pydantic = "^2.9.2"
google-cloud-aiplatform = "^1.71.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
